# openapi.yaml
openapi: 3.0.3
info:
  title: Vision Financial API
  version: "1.0"
  description: >
    Self-hosted financial transaction manager.
    All responses are wrapped in the ADR-026 envelope:
    `{ ok: true, data: T }` on success, `{ ok: false, error: {...} }` on failure.
servers:
  - url: http://localhost:3000
    description: Local development

# ─── Shared envelope & list wrappers ─────────────────────────────────────────
components:
  parameters:
...
    OptionalLimit:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
      description: >
        Page size. Omit (together with offset) to receive the whole collection.
        Values above the endpoint's cap are clamped.
    OptionalOffset:
      in: query
      name: offset
      required: false
      schema:
        type: integer
        minimum: 0
      description: Rows to skip. Omit (together with limit) to receive the whole collection.
  schemas:
...
    Envelope:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        data:
          description: Present on success
        error:
          $ref: "#/components/schemas/ErrorPayload"
    ErrorPayload:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
        status:
          type: integer
    Link:
      type: object
      required: [rel, href, method]
      properties:
        rel:
          type: string
        href:
          type: string
        method:
          type: string
        title:
          type: string

    # Pagination travels in the response BODY, mixed in alongside `items` —
    # never in the envelope's `meta`. `total` is the full match count, not the
    # length of the page.
...
    OptionalPaginationFields:
      type: object
      required: [total]
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # ─── Categories ─────────────────────────────────────────────────────────
...
    Account:
      type: object
      required: [id, name, currency, type, liquidity_class, spendable, in_net_worth, tax_wrapper, owner, multi_currency_cash, has_cash_sleeve, is_active, created_at, updated_at]
      properties:
        id:
          type: integer
        name:
          type: string
        display_name:
          type: string
          nullable: true
        institution:
          type: string
          nullable: true
        currency:
          type: string
          description: ISO 4217 currency code
        type:
          type: string
          enum: [checking, savings, brokerage, crypto_exchange, wallet, pension, liability]
        liquidity_class:
          type: string
          enum: [liquid, semi_liquid, illiquid]
        spendable:
          type: boolean
        in_net_worth:
          type: boolean
        tax_wrapper:
          type: string
          enum: [none, pension, tax_advantaged]
        owner:
          type: string
          enum: [me, partner, joint]
        multi_currency_cash:
          type: boolean
        has_cash_sleeve:
          type: boolean
        funding_account_id:
          type: integer
          nullable: true
        statement_balance:
          type: number
          nullable: true
        statement_balance_date:
          type: string
          format: date
          nullable: true
        computed_balance:
          type: number
          nullable: true
          description: >-
            The account's anchor+delta computed balance (ADR-094), denominated in
            `currency`. On a multi-currency account each currency partition is
            computed separately and converted into `currency` at today's rate, so
            this figure moves with exchange rates.
        reconcilable_balance:
          type: number
          nullable: true
          description: >-
            The reconciliation base: the computed balance of the single currency
            partition `statement_balance` is a statement for, in
            `reconcilable_currency` and NOT FX-converted. Equals
            `computed_balance` for a single-currency account and differs on a
            multi-currency one, so the reconcile dialog previews an entered
            reading against THIS figure — it is what `POST /accounts/{id}/reconcile`
            resolves against. Only returned by the list endpoint.
        reconcilable_currency:
          type: string
          description: >-
            Currency of `reconcilable_balance` and `drift` — normally `currency`,
            but the account's sole funded partition's code when the ledger's
            currency disagrees with the declared one. Only returned by the list
            endpoint.
        drift:
          type: number
          nullable: true
          description: >-
            statement_balance − reconcilable_balance, in `reconcilable_currency`
            (ADR-094); null if no statement balance. Native-currency by design —
            never statement_balance − computed_balance, which on a multi-currency
            account would make the badge move with the daily exchange rate.
        anchor_date:
          type: string
          format: date
          description: YYYY-MM-DD date of the stamped statement balance anchoring computed_balance (WP-A1 provenance); absent when nothing is stamped. Only returned by the list endpoint.
        post_anchor_count:
          type: integer
          description: Active ledger entries after the anchor, or all active entries when unstamped — the "{n} entries since" / "sum of {n} entries" figure (WP-A1 provenance). Only returned by the list endpoint.
        has_transactions:
          type: boolean
          description: Whether the account has any active ledger rows; only returned by the list endpoint.
        is_active:
          type: boolean
        closed_at:
          type: string
          format: date-time
          nullable: true
          description: Server-stamped when the account is closed (is_active=false); cleared on reactivate (ADR-088 addendum, D5)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AccountList:
      allOf:
        - $ref: "#/components/schemas/OptionalPaginationFields"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Account"
            links:
              type: array
              items:
                $ref: "#/components/schemas/Link"
    AccountCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        display_name:
          type: string
          nullable: true
        institution:
          type: string
          nullable: true
        currency:
          type: string
        type:
          type: string
          enum: [checking, savings, brokerage, crypto_exchange, wallet, pension, liability]
        liquidity_class:
          type: string
          enum: [liquid, semi_liquid, illiquid]
        spendable:
          type: boolean
        in_net_worth:
          type: boolean
        tax_wrapper:
          type: string
          enum: [none, pension, tax_advantaged]
        owner:
          type: string
          enum: [me, partner, joint]
        multi_currency_cash:
          type: boolean
        has_cash_sleeve:
          type: boolean
        funding_account_id:
          type: integer
          nullable: true
        statement_balance:
          type: number
          nullable: true
        statement_balance_date:
          type: string
          format: date
          nullable: true
    AccountUpdate:
      type: object
      properties:
        name:
          type: string
        display_name:
          type: string
          nullable: true
        institution:
          type: string
          nullable: true
        currency:
          type: string
        type:
          type: string
          enum: [checking, savings, brokerage, crypto_exchange, wallet, pension, liability]
        liquidity_class:
          type: string
          enum: [liquid, semi_liquid, illiquid]
        spendable:
          type: boolean
        in_net_worth:
          type: boolean
        tax_wrapper:
          type: string
          enum: [none, pension, tax_advantaged]
        owner:
          type: string
          enum: [me, partner, joint]
        multi_currency_cash:
          type: boolean
        has_cash_sleeve:
          type: boolean
        funding_account_id:
          type: integer
          nullable: true
        statement_balance:
          type: number
          nullable: true
        statement_balance_date:
          type: string
          format: date
          nullable: true
        is_active:
          type: boolean
...
paths:
...
  /api/accounts:
    get:
      summary: List accounts
      operationId: getAccounts
      tags: [Accounts]
      parameters:
        - in: query
          name: active
          schema:
            type: string
            enum: ["true", "false", "all"]
          description: Filter by active status (default true)
        - $ref: "#/components/parameters/OptionalLimit"
        - $ref: "#/components/parameters/OptionalOffset"
      responses:
        "200":
          description: Account list (complete unless limit/offset are supplied)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AccountList"
    post:
      summary: Create account
      operationId: createAccount
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountCreate"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Account"
        "400":
          description: Invalid field (e.g. funding_account_id not a positive integer or not an existing account), or statement_balance set without statement_balance_date
        "409":
          description: An account with that name already exists
  /api/accounts/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int32
          minimum: 1
    get:
      summary: Get one account
      operationId: getAccount
      tags: [Accounts]
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Account"
        "404":
          description: Account not found
    patch:
      summary: Update account
      operationId: updateAccount
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Account"
        "400":
          description: Invalid field (e.g. funding_account_id not a positive integer, self-referencing, not an existing account, or closing a funding cycle), or an invalid statement_balance/statement_balance_date pairing
        "404":
          description: Account not found
        "409":
          description: An account with that name already exists
    delete:
      summary: Delete account (409 if still referenced — archive instead)
      operationId: deleteAccount
      tags: [Accounts]
      responses:
        "204":
          description: No Content
        "404":
          description: Account not found
        "409":
          description: Account still has transactions; archive it instead
  /api/accounts/{id}/merge:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int32
          minimum: 1
        description: Survivor account id (sources merge into this one)
    post:
      summary: Merge source accounts into this (survivor) account
      description: Repoints all references (transactions, planned, portfolio lots, funding_account_id) from each source account to this (survivor) account, then deletes the sources (ADR-088).
      operationId: mergeAccounts
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source_ids]
              properties:
                source_ids:
                  type: array
                  minItems: 1
                  maxItems: 500
                  items:
                    type: integer
                    format: int32
                    minimum: 1
                    maximum: 2147483647
                  description: Account ids to merge into this account and delete; must not include the survivor id
      responses:
        "200":
          description: Merge result (survivor id, merged ids, reassigned counts)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "400":
          description: Empty, oversized, malformed, or self-referencing source_ids list
        "404":
          description: Target or a source account not found
  /api/accounts/{id}/merge-preview:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int32
          minimum: 1
        description: Source account id (the account that would be merged away)
    get:
      summary: Preview merging this account into another (read-only)
      description: >-
        Dry-run of merging account {id} (the source) INTO the account given by
        `into` (the survivor). Returns the row counts that would be repointed
        (same categories POST /merge moves), the projected post-merge computed
        balance — the anchor+delta definition evaluated over the union of both
        accounts' active rows, in the survivor's native currency — and whether
        the merge would interleave stamped balance histories
        (`stampsInterleaved`), in which case the merge clears the survivor's
        statement-balance anchor. No mutation.
      operationId: previewAccountMerge
      tags: [Accounts]
      parameters:
        - in: query
          name: into
          required: true
          schema:
            type: integer
          description: Survivor account id (the account {id} would merge into)
      responses:
        "200":
          description: Merge preview
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          into:
                            type: integer
                            description: Survivor account id
                          source:
                            type: integer
                            description: Source account id
                          reassigned:
                            type: object
                            description: Row counts that would move to the survivor
                            properties:
                              transactions: { type: integer }
                              planned: { type: integer }
                              portfolio: { type: integer }
                              funding: { type: integer }
                          projectedBalance:
                            type: number
                            description: Post-merge computed balance — anchor+delta evaluated per currency over the union of both accounts' active rows, each currency partition converted at its own current rate into the survivor's currency
                          projectedBalanceCurrency:
                            type: string
                            description: The survivor's native currency (ISO-4217)
                          stampsInterleaved:
                            type: boolean
                            description: Whether both accounts carry stamped balance histories with overlapping date ranges (the merge would clear the survivor's statement anchor)
                          openingAnchorCollision:
                            type: boolean
                            description: Whether both accounts hold an opening balance in the same currency (only one is allowed per account and currency, so POST /merge would refuse with a 400 until one is removed)
        "400":
          description: Missing/invalid `into`, or `into` equals {id}
        "404":
          description: Source or survivor account not found
  /api/accounts/{id}/opening-balance:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int32
          minimum: 1
        description: Account to anchor
    post:
      summary: Set the opening-balance anchor for an account
      description: >-
        Creates (or updates) one system anchor row per account+currency with
        amount=0, a server-stamped `balance`, is_transfer=true and
        transfer_source='opening' (ADR-094 second addendum, D4). The single
        sanctioned exception to the balance write-protection; lets manual/cash-only
        accounts seed a balance. Invoking again updates the existing anchor.
      operationId: setAccountOpeningBalance
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [balance, date]
              properties:
                balance:
                  type: number
                  description: The opening balance to stamp
                date:
                  type: string
                  format: date
                  description: As-of date; expected to precede the account's activity
                currency:
                  type: string
                  description: 3-letter ISO code; defaults to the account currency
      responses:
        "200":
          description: The anchor transaction plus an optional warning
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "400":
          description: Missing/invalid balance, date, or currency (must be a 3-letter ISO code)
        "404":
          description: Account not found
  /api/accounts/{id}/reconcile:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int32
          minimum: 1
        description: Account to reconcile
    post:
      summary: Reconcile an account's drift
      description: >-
        Resolves the drift (statement_balance − reconcilable_balance) surfaced by
        the drift badge (ADR-094, Phase C). Both modes act on the reconciliation
        base — the one currency partition the statement figure is a statement for
        — so a multi-currency account resolves in its own currency and the other
        partitions are untouched. `mode: 'accept'` rewrites the stored statement
        figure to that base (no transaction created); `mode: 'adjustment'` stamps
        one server-side ledger row (amount=drift, in `reconcilable_currency`,
        balance-free, is_transfer=true, transfer_source='adjustment') so the base
        rises to meet the statement — preserving the ADR-094 descriptive-only
        default. Either way the drift collapses to 0.
      operationId: reconcileAccount
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mode]
              properties:
                mode:
                  type: string
                  enum: [accept, adjustment]
                  description: How to resolve the drift
      responses:
        "200":
          description: The reconciled figures plus any created adjustment transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "400":
          description: No drift to resolve, no statement balance, or an invalid mode
        "404":
          description: Account not found

  # ─── Transactions ────────────────────────────────────────────────────────────
...